fix(lane-create): use the version from .bitmap when creating a lane from lane
